[RENAME] 대소문자 변경을 위한 중간단계 #NADA-49
