+++ src/components/common/modal/PrintCenteredText.js
@@ -1,6 +1,6 @@
 /** PrintCenterText 모달 내부에 가운데정렬 TEXT를 출력하는 컴포넌트 */
 import classNames from 'classnames';
-import { Menu } from '../../../styles/community/BarModalStyle';
+import { Menu } from '../../../styles/mommunity/BarModalStyle';
 
 export const PrintCenteredText = ({ print, act }) => {
   const isRed = print === '삭제하기';
