+++ src/components/community/PostList/PostItem.js
@@ -1,5 +1,5 @@
 /** PostItem, postList을 이루는 PostItem. */
-import { PostContainer } from '../../../styles/community/postItemStyle';
+import { PostContainer } from '../../../styles/mommunity/postItemStyle';
 
 const PostItem = ({ card }) => {
   const { title, field, category, region, content, postAt } = card;
