+++ src/containers/community/PostWriteButton.js
@@ -1,5 +1,5 @@
 /** PostWriteButton 메인페이지에서 작성 페이지로 넘어가는 버튼 컨테이너 */
-import { PostButton } from '../../styles/community/PostWriteButtonStyle';
+import { PostButton } from '../../styles/mommunity/PostWriteButtonStyle';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import { setField } from '../../modules/Community/postWrite';
