+++ src/pages/community/PostWritePage.js
@@ -8,7 +8,7 @@ import {
   postWriteSelector,
 } from '../../modules/Community/postWrite';
 import { setBarStatus } from '../../modules/bar';
-import { PostContainer, Border } from '../../styles/community/PostWriteStyle';
+import { PostContainer, Border } from '../../styles/mommunity/PostWriteStyle';
 import { Title } from '../../containers/community/postWrite/Title';
 import { FilterBar } from '../../containers/community/postWrite/FilterBar';
 import { Content } from '../../containers/community/postWrite/Content';
