--- src/styles/community/PostWriteButtonStyle.js
@@ -1,19 +1,0 @@
-import styled from 'styled-components';
-
-export const PostButton = styled.div`
-  background-color: var(--myspec-primaryblue-1);
-  border-radius: 36px;
-  height: 60px;
-  width: 60px;
-  position: sticky;
-  bottom: 100px;
-  left: 77%;
-
-  & > .group {
-    height: 26px;
-    left: 12px;
-    position: relative;
-    top: 17px;
-    width: 37px;
-  }
-`;
