+++ src/styles/mommunity/CommunityBarStyle.js
