+++ src/styles/mommunity/NoticeModalStyle.js
