+++ src/styles/mommunity/PostDetailStyle.js
